fix(conformance): avoid crash when metadata file is missing

+++ spec/conformance/conformance_cache_store.rb
@@ -8,13 +8,15 @@ class ConformanceCacheStore < FidoMetadata::TestCacheStore
   FILENAME = "metadata.zip"
 
   def setup_authenticators
-    puts("#{FILENAME} not found, this will affect Metadata Service Test results.") unless File.exist?(FILENAME)
-
-    Zip::File.open(FILENAME).glob("metadataStatements/*.json") do |file|
-      json = JSON.parse(file.get_input_stream.read)
-      statement = FidoMetadata::Statement.from_json(json)
-      identifier = statement.aaguid || statement.attestation_certificate_key_identifiers.first
-      write("statement_#{identifier}", statement)
+    if File.exist?(FILENAME)
+      Zip::File.open(FILENAME).glob("metadataStatements/*.json") do |file|
+        json = JSON.parse(file.get_input_stream.read)
+        statement = FidoMetadata::Statement.from_json(json)
+        identifier = statement.aaguid || statement.attestation_certificate_key_identifiers.first
+        write("statement_#{identifier}", statement)
+      end
+    else
+      puts("#{FILENAME} not found, this will affect Metadata Service Test results.")
     end
   end
